Move `BeamSearchScorer` to its own module
