+++ shortfin/python/shortfin_apps/llm/components/token_selection_strategy/beam_group.py
@@ -6,7 +6,7 @@
 
 import logging
 import numpy as np
-from typing import Union, ForwardRef
+from typing import Tuple, Union
 
 import shortfin.array as sfnp
 
@@ -22,7 +22,7 @@ from ..messages import LlmInferenceExecRequest
 
 logger = logging.getLogger(__name__)
 
-BaseBeamScorer = ForwardRef("BaseBeamScorer")
+TOP_P_DEFAULT_SELECTION = 32
 
 
 @dataclass
@@ -195,95 +195,195 @@ class BaseBeam(ABC):
         )
 
 
-class BaseBeamScorer(ABC):
-    def __init__(self, config):
-        self.config = config
+class Beam(BaseBeam):
+    def _sample_greedy(self, logits: np.array, indices: Union[np.array, None]) -> int:
+        """Select the token with the highest logit value.
 
-    @abstractmethod
-    def update_score(
+        Args:
+            logits (np.array): The logits from which to select.
+            indices (np.array | None): Optional indices to filter logits.
+
+        Returns:
+            int: The token ID of the selected token.
+        """
+        if indices is not None:
+            return indices.items[0]
+
+        return self.sampler.select_greedy(logits)
+
+    def _sample_beam_search(
+        self, logits: np.array, indices: Union[np.array, None], k: int
+    ) -> Tuple[np.array, np.array]:
+        """Select the top-k tokens based on beam search.
+
+        Args:
+            logits (np.array): The logits from which to select.
+            indices (np.array | None): Optional indices to filter logits.
+
+        Returns:
+            Tuple[np.array, np.array]: The selected tokens and their probabilities.
+        """
+        if indices is not None:
+            indices = np.array(indices)
+
+        tokens, probs = self.sampler.select_top_k(logits, indices, -k)
+
+        if self.decode_config.logits_normalization == LogitsNormalization.NONE:
+            probs = self.apply_temperature(probs)
+
+        log_probs = self.convert_logits_normalization(
+            self.decode_config.logits_normalization,
+            LogitsNormalization.LOG_SOFTMAX,
+            probs,
+        ).tolist()
+
+        return tokens, log_probs
+
+    def _convert_results_to_log_probs(
         self,
-        beam: BaseBeam,
-        value: float,
-    ) -> None:
-        """Update the score of a `beam`.
+        probs: np.array,
+    ):
+        log_probs = self.convert_logits_normalization(
+            LogitsNormalization.SOFTMAX,
+            LogitsNormalization.LOG_SOFTMAX,
+            probs,
+        )
 
-        Args:
-            beam (BaseBeam): The beam to update.
-            value (float): Value to update the score with.
-        """
+        return log_probs.tolist()
 
-    @abstractmethod
-    def finalize_score(self, beam: BaseBeam) -> None:
-        """Define a `final_score` for a given beam, if applicable.
+    def _pre_select_top_p(
+        self, logits: np.array, indices: Union[np.array, None]
+    ) -> Tuple[np.array, np.array]:
+        top_p_selection = min(logits.shape[-1], TOP_P_DEFAULT_SELECTION)
+        tokens, values = self.sampler.select_top_k(logits, indices, -top_p_selection)
+        probs = self._to_softmax(
+            values,
+            self.decode_config.logits_normalization,
+        )
 
-        Args:
-            beam (BaseBeam): The beam to update.
-        """
+        if indices is None:
+            sorted_order = np.argsort(probs)[::-1]
+            tokens = tokens[sorted_order]
+            probs = probs[sorted_order]
 
-    @abstractmethod
-    def normalize_score(
+        return tokens, probs
+
+    def _sample_default(
+        self, logits: np.array, indices: Union[np.array, None], k: int | None = None
+    ) -> int:
+        decode_config = self.decode_config
+        if decode_config.use_beam_search:
+            return self._sample_beam_search(logits, indices, k)
+        return self._sample_greedy(logits, indices)
+
+    def _sample_top_k(
         self,
-        beam: BaseBeam,
-        value: float,
-    ) -> float:
-        """Normalize the score of a `beam`.
+        logits: np.array,
+        indices: Union[np.array, None],
+        top_k: int,
+    ) -> Tuple[np.array, np.array]:
+        """Sample the top-k tokens from the logits.
 
         Args:
-            beam (BaseBeam): The beam to normalize.
-            value (float): Value to normalize the score with.
+            logits (np.array): The logits from which to select.
+            indices (np.array | None): Optional indices to filter logits.
+            top_k (int): The number of top tokens to select.
+            num_selections (int): The number of selections to make.
 
         Returns:
-            float: Normalized score.
+            Tuple[np.array, np.array]: The selected tokens and their probabilities.
         """
+        decode_config = self.decode_config
+        use_beam_search = decode_config.use_beam_search
+        top_p = decode_config.top_p
 
-    @abstractmethod
-    def score_beams(beams: List[BaseBeam]) -> List[BaseBeam]:
-        """Score a group of beams.
+        num_selections = top_k if use_beam_search or (top_p is not None) else 1
+        return self._sample_logits_top_k(
+            logits,
+            indices,
+            top_k,
+            num_selections,
+        )
+
+    def sample_logits(self, num_completed_beams: int) -> int:
+        """Return the single highest scoring token of the logits.
 
         Args:
-            beams (List[BaseBeam]): The beams to score.
+            num_active_beams (int): Number of active beams, used for beam search.
 
         Returns:
-            List[BaseBeam]: The scored beams in descending order of score.
+            int: The `argmax` of the logits.
         """
+        exec_req = self.exec_req
+        decode_config = self.decode_config
 
-    @abstractmethod
-    def select_beams(
-        active_beams: List[BaseBeam], complete_beams: List[BaseBeam]
-    ) -> List[BaseBeam]:
-        """Select the next candidate set of beams for decode invocation.
+        use_beam_search = decode_config.use_beam_search
+        num_beams = decode_config.num_beams
+        k = num_beams - num_completed_beams
+        top_k = decode_config.top_k
+        top_p = decode_config.top_p
+
+        logits = np.array(exec_req.result_logits)
+        indices = exec_req.result_indices
+
+        if (top_k, top_p) == (None, None):
+            return self._sample_default(logits, indices, k)
+
+        indices = np.array(indices) if indices is not None else None
+        if top_k is not None:
+            tokens, probs = self._sample_top_k(
+                logits,
+                indices,
+                top_k,
+            )
+
+        if top_p is not None:
+            if top_k is None:
+                tokens, probs = self._pre_select_top_p(logits, indices)
+
+            tokens, probs = self._sample_logits_top_p(
+                tokens,
+                probs,
+                top_p,
+                k if use_beam_search else 1,
+                return_probs=use_beam_search,
+            )
+
+        if use_beam_search:
+            log_probs = self._convert_results_to_log_probs(
+                probs,
+            )
+            return tokens, log_probs
+
+        return int(tokens[0])
+
+    def update_exec_req(self):
+        """Update the `LlmInferenceExecRequest` with the selected token."""
+        self.exec_req.input_token_ids.append(self.last_token)
+        self.exec_req.start_position += 1
+
+    def update_score(self, log_prob: float):
+        """Increment the cumulative_log_prob of the beam.
 
         Args:
-            active_beams (List[BaseBeam]): The beams still actively being decoded.
-            complete_beams (List[BaseBeam]): The beams that are completed.
-
-        Returns:
-            List[BaseBeam]: Selected beams.
+            log_prob (float): Log probability of the token.
         """
+        self.score += log_prob
 
-    @abstractmethod
-    def reset(self) -> None:
-        """Reset the state of the scorer.
-
-        This is useful when reusing the scorer for multiple decoding iterations.
-        """
-
-    def penalize_brevity(
-        self,
-        beam: BaseBeam,
-    ) -> float:
-        """Apply a length penalty to the score of a `beam`.
+    def normalize_score(self, min_log_prob: float):
+        """Track the accumulated_normalization for a given beam.
 
         Args:
-            beam (BaseBeam): The beam to penalize.
-            length (int): Length of the sequence.
-
-        Returns:
-            float: Penalized score.
+            min_log_prob (float): Minimum log probability of the selected tokens.
         """
-        # TODO(stbaione): Extend this to support other length penalty types
-        exec_req = beam.exec_req
-        beam.score /= len(exec_req.input_token_ids) - exec_req.prompt_length
+        self.accumulated_normalization += abs(min_log_prob)
+
+    def update_final_score(self):
+        """Calculate the final score of a beam, with a brevity penalty."""
+        exec_req = self.exec_req
+        self.score = (self.score - self.accumulated_normalization) / (
+            len(exec_req.input_token_ids) - exec_req.prompt_length
+        )
 
 
 class BeamGroup:
